Move object provider to its own file
