+++ dsn/src/dsn-object-provider.js
@@ -1,0 +1,23 @@
+import DsnUtils from './DsnUtils.js';
+import dictionary from '../res/dsn-dictionary.json';
+
+export const objectProvider = {
+    get: function (identifier) {
+        if (identifier.key === 'dsn') {
+            return Promise.resolve({
+                identifier: {
+                    namespace: 'deep.space.network',
+                    key: 'dsn'
+                },
+                type: 'folder',
+                location: 'ROOT',
+                name: 'Deep Space Network',
+                composition: []
+            });
+        } else {
+            return Promise.resolve(
+                dictionary.domainObjects[DsnUtils.serializeIdentifier(identifier)]
+            );
+        }
+    }
+};
